requirements.txt
aiohttp==3.9.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
black==23.11.0
celery==5.3.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.5
Django==4.2.7
django-celery-beat==2.5.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
idna==3.4
kombu==5.3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pika==1.3.2
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.41
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
twilio==8.10.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
yarl==1.9.2
